frameworks/Otel-Kieker-Python/requirements.txt
opentelemetry-api==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-instrumentation==0.40b0

